PhrasePage: Replaced pykdeuic4 generated code with cleaned up original ui file. Merged with UI_PhrasePage class and moved everything to phrasepage.py
