+++ lib/autokey/qtui/configwindow.py
@@ -55,130 +55,8 @@ def set_url_label(button, path):   # TODO: phase out in favour of fixed version 
 # ---- Internal widgets
 
 from .scriptpage import ScriptPage
+from .phrasepage import PhrasePage
 
-
-
-from . import phrasepage
-
-class PhrasePage(QWidget, phrasepage.Ui_PhrasePage):
-
-    def __init__(self):
-        QWidget.__init__(self)
-        phrasepage.Ui_PhrasePage.__init__(self)
-        self.setupUi(self)
-
-        self.initialising = True
-        l = list(model.SEND_MODES.keys())
-        l.sort()
-        for val in l:
-            self.sendModeCombo.addItem(val)
-        self.initialising = False
-        
-    def load(self, phrase):
-        self.currentPhrase = phrase
-        self.phraseText.setPlainText(phrase.phrase)
-        self.showInTrayCheckbox.setChecked(phrase.showInTrayMenu)
-
-        for k, v in model.SEND_MODES.items():
-            if v == phrase.sendMode:
-                self.sendModeCombo.setCurrentIndex(self.sendModeCombo.findText(k))
-                break
-                
-        if self.is_new_item():
-            self.urlLabel.setEnabled(False)
-            self.urlLabel.setText(i18n("(Unsaved)"))
-        else:
-            set_url_label(self.urlLabel, self.currentPhrase.path)
-        
-        # TODO - re-enable me if restoring predictive functionality
-        #self.predictCheckbox.setChecked(model.TriggerMode.PREDICTIVE in phrase.modes)
-        
-        self.promptCheckbox.setChecked(phrase.prompt)
-        self.settingsWidget.load(phrase)
-        
-    def save(self):
-        self.settingsWidget.save()
-        self.currentPhrase.phrase = str(self.phraseText.toPlainText())
-        self.currentPhrase.showInTrayMenu = self.showInTrayCheckbox.isChecked()
-
-        self.currentPhrase.sendMode = model.SEND_MODES[str(self.sendModeCombo.currentText())]
-        
-        # TODO - re-enable me if restoring predictive functionality
-        #if self.predictCheckbox.isChecked():
-        #    self.currentPhrase.modes.append(model.TriggerMode.PREDICTIVE)
-        
-        self.currentPhrase.prompt = self.promptCheckbox.isChecked()
-        
-        self.currentPhrase.persist()
-        set_url_label(self.urlLabel, self.currentPhrase.path)
-        return False
-
-    def set_item_title(self, title):
-        self.currentPhrase.description = title
-        
-    def rebuild_item_path(self):
-        self.currentPhrase.rebuild_path()
-        
-    def is_new_item(self):
-        return self.currentPhrase.path is None
-        
-    def reset(self):
-        self.load(self.currentPhrase)   
-        
-    def validate(self):        
-        errors = []
-        
-        # Check phrase content
-        phrase = str(self.phraseText.toPlainText())
-        if EMPTY_FIELD_REGEX.match(phrase):
-            errors.append(i18n("The phrase content can't be empty"))
-            
-        # Check settings
-        errors += self.settingsWidget.validate()
-        
-        if errors:
-            msg = i18n(PROBLEM_MSG_SECONDARY, '\n'.join([str(e) for e in errors]))
-            KMessageBox.detailedError(self.topLevelWidget(), PROBLEM_MSG_PRIMARY.toString(), msg)
-                
-        return len(errors) == 0
-        
-    def set_dirty(self):
-        self.topLevelWidget().set_dirty()
-        
-    def undo(self):
-        self.phraseText.undo()
-        
-    def redo(self):
-        self.phraseText.redo()
-        
-    def insert_token(self, token):
-        self.phraseText.insertPlainText(token)
-        
-    # --- Signal handlers       
-    def on_phraseText_textChanged(self):
-        self.set_dirty()
-        
-    def on_phraseText_undoAvailable(self, state):
-        self.topLevelWidget().set_undo_available(state)
-        
-    def on_phraseText_redoAvailable(self, state):
-        self.topLevelWidget().set_redo_available(state)
-    
-    def on_predictCheckbox_stateChanged(self, state):
-        self.set_dirty()
-        
-    def on_promptCheckbox_stateChanged(self, state):
-        self.set_dirty()
-
-    def on_showInTrayCheckbox_stateChanged(self, state):
-        self.set_dirty()
-
-    def on_sendModeCombo_currentIndexChanged(self, index):
-        if not self.initialising:
-            self.set_dirty()
-    
-    def on_urlLabel_leftClickedUrl(self, url=None):
-        if url: subprocess.Popen(["/usr/bin/xdg-open", url])
 
 
 from . import folderpage
