+++ lib/autokey/qtui/phrasepage.py
@@ -1,73 +1,147 @@
-#!/usr/bin/env python
-# coding=UTF-8
-#
-# Generated by pykdeuic4 from phrasepage.ui on Sun Mar  4 11:39:40 2012
-#
-# WARNING! All changes to this file will be lost.
-from PyKDE4 import kdecore
-from PyKDE4 import kdeui
-from PyQt4 import QtCore, QtGui
+# Copyright (C) 2011 Chris Dekter
+# Copyright (C) 2018 Thomas Hess <thomas.hess@udo.edu>
 
-try:
-    _fromUtf8 = QtCore.QString.fromUtf8
-except AttributeError:
-    _fromUtf8 = lambda s: s
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
 
-class Ui_PhrasePage(object):
-    def setupUi(self, PhrasePage):
-        PhrasePage.setObjectName(_fromUtf8("PhrasePage"))
-        PhrasePage.resize(540, 421)
-        self.verticalLayout_2 = QtGui.QVBoxLayout(PhrasePage)
-        self.verticalLayout_2.setObjectName(_fromUtf8("verticalLayout_2"))
-        self.urlLabel = KUrlLabel(PhrasePage)
-        self.urlLabel.setAlignment(QtCore.Qt.AlignCenter)
-        self.urlLabel.setObjectName(_fromUtf8("urlLabel"))
-        self.verticalLayout_2.addWidget(self.urlLabel)
-        self.phraseText = KTextEdit(PhrasePage)
-        self.phraseText.setTabChangesFocus(True)
-        self.phraseText.setLineWrapMode(QtGui.QTextEdit.NoWrap)
-        self.phraseText.setAcceptRichText(False)
-        self.phraseText.setObjectName(_fromUtf8("phraseText"))
-        self.verticalLayout_2.addWidget(self.phraseText)
-        self.settingsGroupBox = QtGui.QGroupBox(PhrasePage)
-        self.settingsGroupBox.setObjectName(_fromUtf8("settingsGroupBox"))
-        self.verticalLayout = QtGui.QVBoxLayout(self.settingsGroupBox)
-        self.verticalLayout.setObjectName(_fromUtf8("verticalLayout"))
-        self.promptCheckbox = QtGui.QCheckBox(self.settingsGroupBox)
-        self.promptCheckbox.setObjectName(_fromUtf8("promptCheckbox"))
-        self.verticalLayout.addWidget(self.promptCheckbox)
-        self.showInTrayCheckbox = QtGui.QCheckBox(self.settingsGroupBox)
-        self.showInTrayCheckbox.setObjectName(_fromUtf8("showInTrayCheckbox"))
-        self.verticalLayout.addWidget(self.showInTrayCheckbox)
-        self.horizontalLayout_2 = QtGui.QHBoxLayout()
-        self.horizontalLayout_2.setObjectName(_fromUtf8("horizontalLayout_2"))
-        self.label = QtGui.QLabel(self.settingsGroupBox)
-        self.label.setObjectName(_fromUtf8("label"))
-        self.horizontalLayout_2.addWidget(self.label)
-        self.sendModeCombo = QtGui.QComboBox(self.settingsGroupBox)
-        self.sendModeCombo.setObjectName(_fromUtf8("sendModeCombo"))
-        self.horizontalLayout_2.addWidget(self.sendModeCombo)
-        spacerItem = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.horizontalLayout_2.addItem(spacerItem)
-        self.verticalLayout.addLayout(self.horizontalLayout_2)
-        self.kseparator = KSeparator(self.settingsGroupBox)
-        self.kseparator.setObjectName(_fromUtf8("kseparator"))
-        self.verticalLayout.addWidget(self.kseparator)
-        self.settingsWidget = SettingsWidget(self.settingsGroupBox)
-        self.settingsWidget.setObjectName(_fromUtf8("settingsWidget"))
-        self.verticalLayout.addWidget(self.settingsWidget)
-        self.verticalLayout_2.addWidget(self.settingsGroupBox)
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
 
-        self.retranslateUi(PhrasePage)
-        QtCore.QMetaObject.connectSlotsByName(PhrasePage)
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+import subprocess
 
-    def retranslateUi(self, PhrasePage):
-        PhrasePage.setWindowTitle(kdecore.i18n(_fromUtf8("Form")))
-        self.urlLabel.setTipText(kdecore.i18n(_fromUtf8("Open the phrase in the default text editor")))
-        self.settingsGroupBox.setTitle(kdecore.i18n(_fromUtf8("Phrase Settings")))
-        self.promptCheckbox.setText(kdecore.i18n(_fromUtf8("Always prompt before pasting this phrase")))
-        self.showInTrayCheckbox.setText(kdecore.i18n(_fromUtf8("Show in notification icon menu")))
-        self.label.setText(kdecore.i18n(_fromUtf8("Paste using")))
+from PyQt4.QtGui import QMessageBox
 
-from PyKDE4.kdeui import KUrlLabel, KSeparator, KTextEdit
-from .settingswidget import SettingsWidget
+from .common import set_url_label, EMPTY_FIELD_REGEX, inherits_from_ui_file_with_name
+
+from .. import model
+
+PROBLEM_MSG_PRIMARY = "Some problems were found"
+PROBLEM_MSG_SECONDARY = "{}\n\nYour changes have not been saved."
+
+
+# TODO: Once the port to Qt5 is done, set the editor placeholder text in the UI file to "Enter your phrase here."
+# TODO: Pure Qt4 QTextEdit does not support placeholder texts, so this functionality is currently unavailable.
+class PhrasePage(*inherits_from_ui_file_with_name("phrasepage")):
+
+    def __init__(self):
+        super(PhrasePage, self).__init__()
+        self.setupUi(self)
+
+        self.initialising = True
+        l = list(model.SEND_MODES.keys())
+        l.sort()
+        for val in l:
+            self.sendModeCombo.addItem(val)
+        self.initialising = False
+
+    def load(self, phrase):
+        self.currentPhrase = phrase
+        self.phraseText.setPlainText(phrase.phrase)
+        self.showInTrayCheckbox.setChecked(phrase.showInTrayMenu)
+
+        for k, v in model.SEND_MODES.items():
+            if v == phrase.sendMode:
+                self.sendModeCombo.setCurrentIndex(self.sendModeCombo.findText(k))
+                break
+
+        if self.is_new_item():
+            self.urlLabel.setEnabled(False)
+            self.urlLabel.setText("(Unsaved)")  # TODO: i18n
+        else:
+            set_url_label(self.urlLabel, self.currentPhrase.path)
+
+        # TODO - re-enable me if restoring predictive functionality
+        #self.predictCheckbox.setChecked(model.TriggerMode.PREDICTIVE in phrase.modes)
+
+        self.promptCheckbox.setChecked(phrase.prompt)
+        self.settingsWidget.load(phrase)
+
+    def save(self):
+        self.settingsWidget.save()
+        self.currentPhrase.phrase = str(self.phraseText.toPlainText())
+        self.currentPhrase.showInTrayMenu = self.showInTrayCheckbox.isChecked()
+
+        self.currentPhrase.sendMode = model.SEND_MODES[str(self.sendModeCombo.currentText())]
+
+        # TODO - re-enable me if restoring predictive functionality
+        #if self.predictCheckbox.isChecked():
+        #    self.currentPhrase.modes.append(model.TriggerMode.PREDICTIVE)
+
+        self.currentPhrase.prompt = self.promptCheckbox.isChecked()
+
+        self.currentPhrase.persist()
+        set_url_label(self.urlLabel, self.currentPhrase.path)
+        return False
+
+    def set_item_title(self, title):
+        self.currentPhrase.description = title
+
+    def rebuild_item_path(self):
+        self.currentPhrase.rebuild_path()
+
+    def is_new_item(self):
+        return self.currentPhrase.path is None
+
+    def reset(self):
+        self.load(self.currentPhrase)
+
+    def validate(self):
+        errors = []
+
+        # Check phrase content
+        phrase = str(self.phraseText.toPlainText())
+        if EMPTY_FIELD_REGEX.match(phrase):
+            errors.append("The phrase content can't be empty")  # TODO: i18n
+
+        # Check settings
+        errors += self.settingsWidget.validate()
+
+        if errors:
+            msg = PROBLEM_MSG_SECONDARY.format('\n'.join([str(e) for e in errors]))
+            QMessageBox.critical(self.topLevelWidget(), PROBLEM_MSG_PRIMARY, msg)
+
+        return not bool(errors)
+
+    def set_dirty(self):
+        self.topLevelWidget().set_dirty()
+
+    def undo(self):
+        self.phraseText.undo()
+
+    def redo(self):
+        self.phraseText.redo()
+
+    def insert_token(self, token):
+        self.phraseText.insertPlainText(token)
+
+    # --- Signal handlers
+    def on_phraseText_textChanged(self):
+        self.set_dirty()
+
+    def on_phraseText_undoAvailable(self, state):
+        self.topLevelWidget().set_undo_available(state)
+
+    def on_phraseText_redoAvailable(self, state):
+        self.topLevelWidget().set_redo_available(state)
+
+    def on_predictCheckbox_stateChanged(self, state):
+        self.set_dirty()
+
+    def on_promptCheckbox_stateChanged(self, state):
+        self.set_dirty()
+
+    def on_showInTrayCheckbox_stateChanged(self, state):
+        self.set_dirty()
+
+    def on_sendModeCombo_currentIndexChanged(self, index):
+        if not self.initialising:
+            self.set_dirty()
+
+    def on_urlLabel_leftClickedUrl(self, url=None):
+        if url: subprocess.Popen(["/usr/bin/xdg-open", url])
